refactor: removed global token state
I apologize in advance for this dense commit. I took it too far while experimenting.
The main things happening here include:
1. Creating a TokenList struct to compose the list of tokens, the current position in the list, and the number of tokens
(size) in the list
2. Passing pointers of type TokenList to functions involved in parsing and tokenization and using this to manipulate
the state of the token list, thereby removing the need for the global tokens state. This makes it easier to understand
what the code is doing
3. Moved token-specific code to token.h, again spearating concerns, and making things easier to read
I also changed the case of struct members to be in snake case instead of camel case